server/requirements.txt
django>=5.2.6
djangorestframework>=3.16.1
azure-cosmos>=4.5.1
openai>=1.99.0
python-dotenv>=1.0.0
numpy>=1.24.0